WEB_APP/backend/requirements.txt
# AI 網路運維助理 WEB 後端 - 依賴套件清單  
# 2025-08-22 - 移除未使用的依賴

# =============================================================================
# FastAPI 核心框架
# =============================================================================
fastapi[standard]>=0.115.0,<0.120.0
uvicorn[standard]>=0.35.0,<1.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.10.0,<3.0.0

# =============================================================================
# AI 和 LangChain
# =============================================================================
# LangChain 核心套件 - 統一版本管理
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-community>=0.3.0,<0.4.0

# AI 提供者整合
langchain-google-genai>=2.0.0,<3.0.0
langchain-anthropic>=0.3.0,<0.4.0

# Google AI API
google-generativeai>=0.8.0,<1.0.0
google-api-core>=2.19.0,<3.0.0

# Anthropic Claude AI
anthropic>=0.63.0,<1.0.0

# =============================================================================
# 網路自動化和 SSH
# =============================================================================
# SSH 連線套件
asyncssh>=2.17.0,<3.0.0

# =============================================================================
# 配置和環境管理
# =============================================================================
python-dotenv>=1.1.0,<2.0.0
PyYAML>=6.0.1,<7.0.0

# =============================================================================
# 提示詞模板引擎
# =============================================================================
Jinja2>=3.1.0,<4.0.0

# =============================================================================
# 工具套件和增強功能
# =============================================================================
# 重試和錯誤處理
tenacity>=8.3.0,<9.0.0

# =============================================================================
# 系統監控和效能優化
# =============================================================================
# JSON 序列化優化
orjson>=3.11.0,<4.0.0

# =============================================================================
# HTTP 客戶端 - webhook 回調支援 (已移除 httpx，統一使用 aiohttp)
# =============================================================================


# =============================================================================
# 異步 HTTP 客戶端 - webhook 需要 (保留)
# =============================================================================
aiohttp>=3.9.0,<4.0.0

# =============================================================================
# ClickHouse 資料庫客戶端 - Akvorado 流量資料存取
# =============================================================================
clickhouse-connect>=0.7.0,<1.0.0